server/requirements/base.txt
Django==2.2.6
django-modeltranslation==0.13.3
psycopg2-binary==2.8.4
django-bootstrap4==1.0.1
dj-database-url==0.5.0
python-dateutil==2.6.1
python-telegram-bot==9.0.0
ms-api==0.6.58

pymorphy2==0.8
pymorphy2-dicts-ru==2.4.404381.4453942

# for search
django-haystack==2.8.1
Whoosh==2.7.4

# for web parsing
beautifulsoup4==4.6.0
requests==2.18.4

# for monitoring
raven==6.4.0
